// src/api/models/UserVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * 用户信息
 */
export type UserVO = {
  /**
   * 用户 id
   */
  id?: number;
  /**
   * 账号
   */
  userAccount?: string;
  /**
   * 昵称
   */
  userName?: string;
  /**
   * 头像 URL
   */
  userAvatar?: string;
  /**
   * 用户简介
   */
  userProfile?: string;
  /**
   * 角色
   */
  userRole?: string;
  /**
   * 创建时间
   */
  createTime?: string;
  /**
   * 更新时间
   */
  updateTime?: string;
};